refactor: utility function is custom files
