+++ yarn-project/merkle-tree/src/test/utils/create_mem_down.ts
@@ -1,0 +1,3 @@
+import { default as memdown, type MemDown } from 'memdown';
+
+export const createMemDown = () => (memdown as any)() as MemDown<any, any>;
